requirements.txt
flask == 3.0.3
argon2-cffi == 23.1.0
Flask-SQLAlchemy == 3.1.1
psycopg2-binary == 2.9.9
pyjwt == 2.9.0
python-dotenv == 1.0.1